fix(installer): temporarily disable CUDA 13.0 handling
- Add temporary fix to bypass CUDA 13.0 detection in multiple installers
- Set status to "None" when CUDA 13.0 is detected
- Affects comfyui, fooocus, invokeai, sd_trainer, and stable diffusion webui installers

feat(notebook): improve HDM training manager functionality

- Refactor imports to use updated module structure
- Replace direct class usage with functional calls for mirror and environment setup
- Add helper functions for displaying directories and importing Kaggle input
- Simplify repository upload parameters with predefined configs
- Update Git wrapper usage for better integration
- Remove redundant tqdm dependency in installation process

chore(core): update sd-webui-all-in-one module branch

- Change default branch from dev to main for sd-webui-all-in-one installation
- Ensure stable version is used for core module initialization

refactor(manager): move utility functions to dedicated tools

- Move import_kaggle_input and display_model_and_dataset_dir to kaggle_tools
- Remove duplicated methods from SDScriptsManager
- Maintain backward compatibility with existing copy_files method
